+++ tests/configs_analysis/foo_mapping.php
@@ -1,0 +1,20 @@
+<?php
+
+/*
+ * This file is part of the jolicode/elastically library.
+ *
+ * (c) JoliCode <coucou@jolicode.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+return [
+    'mappings' => [
+        'properties' => [
+            'name' => [
+                'type' => 'keyword',
+            ],
+        ],
+    ],
+];