+++ CHANGELOG.md
@@ -5,9 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
 
 ### Fixed
+
+### Added
+
+### Deprecated
+
+## [1.4.0] - 2021-11-10
+
+### Fixed
+
+- Better support for prefixed names
 
 ### Added
 
@@ -126,7 +137,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Allow to set Serializer Context for input/output.
 
-[Unreleased]: https://github.com/jolicode/elastically/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/jolicode/elastically/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/jolicode/elastically/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/jolicode/elastically/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/jolicode/elastically/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/jolicode/elastically/compare/v1.1.0...v1.1.1